feat: add ledger entry creation presets and refactor related functions
